+++ src/commonMain/kotlin/dev/jamesyox/kastro/luna/LunarHorizonEventSequence.kt
@@ -25,13 +25,17 @@ import dev.jamesyox.kastro.util.latitude
 import dev.jamesyox.kastro.util.longitude
 import kotlinx.datetime.Instant
 import kotlin.math.ceil
+import kotlin.math.floor
 import kotlin.time.Duration
 import kotlin.time.Duration.Companion.days
+import kotlin.time.Duration.Companion.hours
 
 /**
  * [Sequence] representing all [LunarEvent.HorizonEvent]s supported by Kastro. The sequence will be ordered by
- * [LunarEvent.time] with times farthest in the future being later in the [Sequence]. As implied by being a [Sequence]
- * values will be lazily calculated so setting a large or even infinite limit is fine.
+ * [LunarEvent.time] either advancing forwards or backwards in time depending on the `reverse` parameter.
+ *
+ * As implied by being a [Sequence] values will be lazily calculated so setting a large or even infinite
+ * limit is acceptable.
  *
  * @param start Time representing the beginning of the sequence
  * @param latitude Latitude of the observer in degrees
@@ -41,14 +45,24 @@ import kotlin.time.Duration.Companion.days
  * Limit only exists to allow certain logical checks such as if you wanted to know if there will be any event in the
  * next hour you could use limit and then check if the resulting sequence is empty. Defaults to 365 days.
  * @param requestedHorizonEvents A [List] of [LunarEvent.HorizonEvent.HorizonEventType]s to calculate event times for.
+ * @param reverse Whether the sequence should advance in reverse chronological order. By default, this is false
  */
 public class LunarHorizonEventSequence(
     private val start: Instant,
     latitude: Double,
     longitude: Double,
     private val limit: Duration = 365.days,
-    private val requestedHorizonEvents: List<LunarEvent.HorizonEvent.HorizonEventType> = LunarEvent.HorizonEvent.all
+    private val requestedHorizonEvents: List<LunarEvent.HorizonEvent.HorizonEventType> = LunarEvent.HorizonEvent.all,
+    private val reverse: Boolean = false,
 ) : Sequence<LunarEvent.HorizonEvent> {
+    @Deprecated(message = "", level = DeprecationLevel.HIDDEN)
+    public constructor(
+        start: Instant,
+        latitude: Double,
+        longitude: Double,
+        limit: Duration = 365.days,
+        requestedHorizonEvents: List<LunarEvent.HorizonEvent.HorizonEventType> = LunarEvent.HorizonEvent.all,
+    ) : this(start, latitude, longitude, limit, requestedHorizonEvents, false)
 
     /**
      * Alternative constructor with location as a [Pair]. Provided as a convenience. See primary constructor for more
@@ -60,14 +74,24 @@ public class LunarHorizonEventSequence(
      * @param limit The limit or outer bound of the Sequence see primary constructor for more information.
      * @param requestedHorizonEvents A [List] of [LunarEvent.HorizonEvent.HorizonEventType]s to calculate event times
      * for.
+     * @param reverse Whether the sequence should advance in reverse chronological order. By default, this is false
      * @see LunarHorizonEventSequence
      */
     public constructor(
         start: Instant,
         location: Pair<Double, Double>,
         limit: Duration = 365.days,
-        requestedHorizonEvents: List<LunarEvent.HorizonEvent.HorizonEventType> = LunarEvent.HorizonEvent.all
-    ) : this(start, location.first, location.second, limit, requestedHorizonEvents)
+        requestedHorizonEvents: List<LunarEvent.HorizonEvent.HorizonEventType> = LunarEvent.HorizonEvent.all,
+        reverse: Boolean = false,
+    ) : this(start, location.first, location.second, limit, requestedHorizonEvents, reverse)
+
+    @Deprecated(message = "", level = DeprecationLevel.HIDDEN)
+    public constructor(
+        start: Instant,
+        location: Pair<Double, Double>,
+        limit: Duration = 365.days,
+        requestedHorizonEvents: List<LunarEvent.HorizonEvent.HorizonEventType> = LunarEvent.HorizonEvent.all,
+    ) : this(start, location.first, location.second, limit, requestedHorizonEvents, false)
 
     private val refraction = apparentRefraction(0.0) // TODO [ALPHA]
     private val lat = latitude.latitude
@@ -81,7 +105,7 @@ public class LunarHorizonEventSequence(
         return generateSequence(execute()) {
             // Take the last of the sequence returned by execute and execute upon it if it is not null.
             execute(
-                startHour = it.hour + 1,
+                startHour = it.hour + (if (reverse) -1 else 1),
                 mode = if (it is InternalLunarHorizonEvent.RiseResult) SearchMode.SET else SearchMode.RISE
             )
         }.map {
@@ -103,7 +127,10 @@ public class LunarHorizonEventSequence(
         val set = execute(startHour = 0, mode = SearchMode.SET)
 
         return if (rise != null && set != null) {
-            if (rise.hour < set.hour) rise else set
+            when (reverse) {
+                true -> if (rise.hour < set.hour) set else rise
+                false -> if (rise.hour < set.hour) rise else set
+            }
         } else {
             rise ?: set
         }
@@ -128,42 +155,56 @@ public class LunarHorizonEventSequence(
         val julianDate = start.julianDate
         var ye: Double
         var hour = startHour
-        val limitHours = limit.inWholeMilliseconds / (60.0 * 60.0 * 1000.0)
-        val maxHours = ceil(limitHours).toInt()
+        val internalLimit = if (reverse) limit * -1 else limit
+        val limitHours = (internalLimit.inWholeMilliseconds.toDouble() / 1.hours.inWholeMilliseconds.toDouble())
+        val maxHours = if (reverse) floor(limitHours).toInt() else ceil(limitHours).toInt()
         var yMinus = correctedLunarHeight(julianDate.atHour(hour - 1.0))
         var y0 = correctedLunarHeight(julianDate.atHour(hour.toDouble()))
         var yPlus = correctedLunarHeight(julianDate.atHour(hour + 1.0))
-        while (hour <= maxHours) {
+        while (if (reverse) hour >= maxHours else hour <= maxHours) {
             val qi = QuadraticInterpolation.of(yMinus, y0, yPlus)
             ye = qi.ye
             if (qi.numberOfRoots == 1) {
                 val rt = qi.root1 + hour
+                val rtPastLimit = if (reverse) (rt <= 0.0 && rt > limitHours) else (rt >= 0.0 && rt < limitHours)
                 if (yMinus < 0.0) {
-                    if (rt >= 0.0 && rt < limitHours && mode.checkRise) {
+                    if (rtPastLimit && mode.checkRise) {
                         return InternalLunarHorizonEvent.RiseResult(jd = julianDate, rt = rt, hour = hour)
                     }
                 } else {
-                    if (rt >= 0.0 && rt < limitHours && mode.checkSet) {
+                    if (rtPastLimit && mode.checkSet) {
                         return InternalLunarHorizonEvent.SetResult(jd = julianDate, rt = rt, hour = hour)
                     }
                 }
             } else if (qi.numberOfRoots == 2) {
                 (hour + if (ye < 0.0) qi.root2 else qi.root1).let { rt ->
-                    if (rt >= 0.0 && rt < limitHours && mode.checkRise) {
+                    val rtPastLimit = if (reverse) (rt <= 0.0 && rt > limitHours) else (rt >= 0.0 && rt < limitHours)
+                    if (rtPastLimit && mode.checkRise) {
                         return InternalLunarHorizonEvent.RiseResult(jd = julianDate, rt = rt, hour = hour)
                     }
                 }
 
                 (hour + if (ye < 0.0) qi.root1 else qi.root2).let { rt ->
-                    if (rt >= 0.0 && rt < limitHours && mode.checkSet) {
+                    val rtPastLimit = if (reverse) (rt <= 0.0 && rt > limitHours) else (rt >= 0.0 && rt < limitHours)
+                    if (rtPastLimit && mode.checkSet) {
                         return InternalLunarHorizonEvent.SetResult(jd = julianDate, rt = rt, hour = hour)
                     }
                 }
             }
-            hour++
-            yMinus = y0
-            y0 = yPlus
-            yPlus = correctedLunarHeight(julianDate.atHour(hour + 1.0))
+            when (reverse) {
+                true -> {
+                    hour--
+                    yPlus = y0
+                    y0 = yMinus
+                    yMinus = correctedLunarHeight(julianDate.atHour(hour - 1.0))
+                }
+                false -> {
+                    hour++
+                    yMinus = y0
+                    y0 = yPlus
+                    yPlus = correctedLunarHeight(julianDate.atHour(hour + 1.0))
+                }
+            }
         }
         return null
     }