+++ src/commonMain/kotlin/dev/jamesyox/kastro/luna/LunarPhaseSequence.kt
@@ -22,6 +22,7 @@ import dev.jamesyox.kastro.util.daysInJulianCentury
 import dev.jamesyox.kastro.util.instant
 import dev.jamesyox.kastro.util.julianDate
 import dev.jamesyox.kastro.util.radians
+import dev.jamesyox.kastro.util.sortedByReversible
 import kotlinx.datetime.Instant
 import kotlin.math.PI
 import kotlin.time.Duration
@@ -31,7 +32,9 @@ import kotlin.time.Duration.Companion.seconds
 /**
  * [Sequence] representing all [LunarEvent.PhaseEvent]s supported by Kastro. The sequence will be ordered by
  * [LunarEvent.time] with times farthest in the future being later in the [Sequence]. As implied by being a [Sequence]
- * values will be lazily calculated so setting a large or even infinite limit is fine.
+ *
+ * As implied by being a [Sequence] values will be lazily calculated so setting a large or even infinite
+ * limit is acceptable.
  *
  * @param start Time representing the beginning of the sequence
  * @param limit Point at which the sequence will end. If you wish for an infinite sequence simply put
@@ -40,12 +43,22 @@ import kotlin.time.Duration.Companion.seconds
  * next hour you could use limit and then check if the resulting sequence is empty. Defaults to 365 days.
  * @param requestedLunarPhases List of [LunarPhase.Primary] events to be included in the sequence. Providing only
  * the phases you are interested in can save computation resources. Defaults to [LunarPhase.Primary.all].
+ * @param reverse Whether the sequence should advance in reverse chronological order. By default, this is false
  */
 public class LunarPhaseSequence(
     private val start: Instant,
     private val limit: Duration = 365.days,
     private val requestedLunarPhases: List<LunarPhase.Primary> = LunarPhase.Primary.all,
+    private val reverse: Boolean = false,
 ) : Sequence<LunarEvent.PhaseEvent> {
+
+    @Deprecated(message = "", level = DeprecationLevel.HIDDEN)
+    public constructor(
+        start: Instant,
+        limit: Duration = 365.days,
+        requestedLunarPhases: List<LunarPhase.Primary> = LunarPhase.Primary.all,
+    ) : this(start, limit, requestedLunarPhases, false)
+
     private companion object {
         private const val SUN_LIGHT_TIME_TAU = 8.32 / (1440.0 * daysInJulianCentury)
     }
@@ -53,7 +66,7 @@ public class LunarPhaseSequence(
     public override fun iterator(): Iterator<LunarEvent.PhaseEvent> {
         return generateSequence(calculateNextLunarPhaseEvents(start)) { moonPhaseEvents ->
             moonPhaseEvents.lastOrNull()?.let { last ->
-                calculateNextLunarPhaseEvents(startingAt = last.time + 1.seconds)
+                calculateNextLunarPhaseEvents(startingAt = last.time + if (reverse) (-500).seconds else 500.seconds)
             }
         }.flatMap { it.asSequence() }.iterator()
     }
@@ -61,24 +74,35 @@ public class LunarPhaseSequence(
     private fun calculateNextLunarPhaseEvents(startingAt: Instant): List<LunarEvent.PhaseEvent> {
         return requestedLunarPhases.mapNotNull {
             calculateNextLunarPhaseEvent(start = startingAt, phase = it)
-        }.sortedBy { it.time }
+        }.sortedByReversible(reverse) { it.time }
     }
 
     private fun calculateNextLunarPhaseEvent(
         start: Instant,
         phase: LunarPhase.Primary,
     ): LunarEvent.PhaseEvent? {
-        val dT = 7.0 / daysInJulianCentury // step rate: 1 week
+        val stepRate = 7.days
+        val dT = stepRate.inWholeDays / daysInJulianCentury
         val accuracy = 0.5 / 1440.0 / daysInJulianCentury // accuracy: 30 seconds
-        var t0 = start.julianDate.julianCentury
+        var t0 = (if (reverse) (start - stepRate) else start).julianDate.julianCentury
         var t1 = t0 + dT
         var d0 = lunarPhase(t0, phase.phase.radians)
         var d1 = lunarPhase(t1, phase.phase.radians)
         while (d0 * d1 > 0.0 || d1 < d0) {
-            t0 = t1
-            d0 = d1
-            t1 += dT
-            d1 = lunarPhase(t1, phase.phase.radians)
+            when (reverse) {
+                true -> {
+                    t1 = t0
+                    d1 = d0
+                    t0 -= dT
+                    d0 = lunarPhase(t0, phase.phase.radians)
+                }
+                false -> {
+                    t0 = t1
+                    d0 = d1
+                    t1 += dT
+                    d1 = lunarPhase(t1, phase.phase.radians)
+                }
+            }
         }
         val tPhase = calculate(t0, t1, accuracy) { lunarPhase(it, phase.phase.radians) }
         val tjd = atJulianCentury(tPhase)