+++ src/commonTest/kotlin/dev/jamesyox/kastro/util/SequenceExtTest.kt
@@ -24,7 +24,22 @@ class SequenceExtTest {
         val sequence2 = sequenceOf(10,20,30,40,50,60,70)
 
         val expected = listOf(1,2,3,4,4,5,6,7,8,10,15,20,30,40,50,60,70)
-        assertEquals(expected , sequence1.mergeWith(sequence2).toList())
+        assertEquals(
+            expected = expected,
+            actual = sequence1.mergeWith(other = sequence2, reverse = false).toList()
+        )
+    }
+
+    @Test
+    fun TestMergeComparableSequencesReverse() {
+        val sequence1 = sequenceOf(15,8,7,6,5,4,4,3,2,1)
+        val sequence2 = sequenceOf(70,60,50,40,30,20,10)
+
+        val expected = listOf(70,60,50,40,30,20,15,10,8,7,6,5,4,4,3,2,1)
+        assertEquals(
+            expected = expected,
+            actual = sequence1.mergeWith(other = sequence2, reverse = true).toList()
+        )
     }
 
     @Test
@@ -33,7 +48,22 @@ class SequenceExtTest {
         val sequence2 = sequenceOf(10,20,30,40,50,60,70)
 
         val expected = listOf(10,20,30,40,50,60,70)
-        assertEquals(expected , sequence1.mergeWith(sequence2).toList())
+        assertEquals(
+            expected = expected,
+            actual = sequence1.mergeWith(sequence2, reverse = false).toList()
+        )
+    }
+
+    @Test
+    fun TestMergeSequencesWhereFirstIsEmptyReverse() {
+        val sequence1 = sequenceOf<Int>()
+        val sequence2 = sequenceOf(70,60,50,40,30,20,10)
+
+        val expected = listOf(70,60,50,40,30,20,10)
+        assertEquals(
+            expected = expected,
+            actual = sequence1.mergeWith(sequence2, reverse = true).toList()
+        )
     }
 
     @Test
@@ -42,6 +72,21 @@ class SequenceExtTest {
         val sequence2 = sequenceOf<Int>()
 
         val expected = listOf(10,20,30,40,50,60,70)
-        assertEquals(expected , sequence1.mergeWith(sequence2).toList())
+        assertEquals(
+            expected = expected,
+            actual = sequence1.mergeWith(sequence2, reverse = false).toList()
+        )
+    }
+
+    @Test
+    fun TestMergeSequencesWhereSecondIsEmptyReverse() {
+        val sequence1 = sequenceOf(70,60,50,40,30,20,10)
+        val sequence2 = sequenceOf<Int>()
+
+        val expected = listOf(70,60,50,40,30,20,10)
+        assertEquals(
+            expected = expected,
+            actual = sequence1.mergeWith(sequence2, reverse = true).toList()
+        )
     }
 }